pyproject.toml
[project]
name = "displacement-tracker"
version = "0.1.0"
description = ""
authors = [{name = "Karim Alaa El-Din", email = "karim@aedin.dev"}]
readme = "README.md"
packages = [{ include = "displacement_tracker" }]

[tool.poetry.dependencies]
python = "^3.10,<3.15"

# --- Core libraries ---
selenium = "^4.33.0"
opencv-python = "^4.8.0"
pillow = "^10.0.0"
numpy = "^1.24.0"
python-dotenv = "^1.1.1"
matplotlib = "^3.10.3"
click = "^8.2.1"
tqdm = "^4.66.1"
google-api-python-client = "^2.176.0"
scipy = "1.15.3"
h5py = "^3.11.0"
PyYAML = "^6.0.1"

# --- GIS libraries (with platform awareness) ---
# These can be difficult to compile on M1 macOS. The markers below ensure pip uses prebuilt wheels.

rasterio = [
    { version = "^1.4.3", markers = "platform_system != 'Darwin'" },
    { version = "^1.4.3", markers = "platform_system == 'Darwin'", extras = ["gdal"] }
]
fiona = [
    { version = "^1.10.1", markers = "platform_system != 'Darwin'" },
    { version = "^1.10.1", markers = "platform_system == 'Darwin'", extras = ["gdal"] }
]
deprecated = "^1.3.1"
torch = "^2.9.1"
torchvision = "^0.24.1"
scikit-learn = "^1.3"
geopandas = "^1.1.2"
psutil = "^7.2.2"
shapely = "^2.0.7"
pyproj = "^3.7.0"

[tool.poetry.scripts]
tif-loader = "displacement_tracker.a_tif_loader:tif_loader"
coordinate-scanner = "displacement_tracker.b_coordinate_scanner:cli"
rebalance-dataset = "displacement_tracker.c_rebalance_dataset:cli"
train-cnn = "displacement_tracker.d_train_cnn:cli"
predict-json = "displacement_tracker.e_predict_json:cli"
evaluate-geojson = "displacement_tracker.f_evaluate_geojson:cli"
merge-geojsons = "displacement_tracker.merge_geojsons:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)"
]
